separate spells + unlock_spell fix
